# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/registration": {
      "post": {
        "operationId": "AuthController_registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TUserRegistration" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TAuthResponseData" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TErrorResponseBody" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TErrorResponseBody" }
              }
            }
          }
        },
        "tags": ["Authorization Controller"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TLoginData" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TAuthResponseData" }
              }
            }
          }
        },
        "tags": ["Authorization Controller"]
      }
    }
  },
  "info": {
    "title": "WBChat-api",
    "description": "Documentation",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "TUserRegistration": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "example@example.com" },
          "first_name": { "type": "string", "example": "Peter" },
          "last_name": { "type": "string", "example": "Bowers" },
          "level": {
            "type": "string",
            "example": "Middle",
            "enum": [
              "Trainee",
              "Junior",
              "Middle",
              "Senior",
              "Tech architector"
            ]
          },
          "direction": {
            "type": "string",
            "example": "Frontend",
            "enum": ["Frontend", "Backend"]
          },
          "experience": {
            "type": "string",
            "example": "No experience",
            "enum": [
              "No experience",
              "Less than 1 year",
              "1-3 years",
              "More than 3 years"
            ]
          },
          "password": { "type": "string" }
        },
        "required": [
          "email",
          "first_name",
          "last_name",
          "level",
          "direction",
          "experience",
          "password"
        ]
      },
      "TAuthResponseData": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGVAZXhhbXBsZS5jb20iLCJpYXQiOjE2NzcyNjc4NDcsImV4cCI6MTY3NzI2OTA0N30.jak0XNbOfp6nNVbaOQONM9aoH9mJ2LlNpDHh3n_1_oY"
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGVAZXhhbXBsZS5jb20iLCJpYXQiOjE2NzcyNjc4NDcsImV4cCI6MTY3NzI2OTA0N30.jak0XNbOfp6nNVbaOQONM9aoH9mJ2LlNpDHh3n_1_oY"
          }
        },
        "required": ["access_token", "refresh_token"]
      },
      "TErrorResponseBody": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Error message" }
        },
        "required": ["message"]
      },
      "TLoginData": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "example@example.com" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      }
    }
  }
}